Show floating point values with "linked precision".
In e.g. a list of float values, it usually doesn't make much sense
to display close-to-zero values with the same number of digits as
larger values, because non-trivial computations will generally smear
the errors from the big values over to the small ones, so we don't
benefit from the in principle smaller epsilons that the small values
could have.