Add (currently failing) test for upwards orders-of-magnitude showing.
We should never change digits, only omit them. Currently, integers in low-precision
contexts get rounded to tens/hundreds/thousands, even when that means actual zero
digits are replaced for the non-significant decimals. This wrongly suggests a higher
precision than is actually there, and then shows digits that are wrong in that
precision.